Fix Swagger problems and add custom location
Previously the precise location was transparent to the app requiring
some funky HTTP headers. Transparent location did not work for Swagger,
so not we configure the location as an environment variable.
Note that the current hosted location (/rpki-history/api) is hard coded
in the openapi.yaml file, so if it is changed in the environment, the
interactive trial feature of the Swagger UI will not work.